[miner] Count `Transaction` bytes directly

+++ src/Chainweb/BlockHeader.hs
@@ -673,7 +673,7 @@ instance FromJSON (ObjectEncoded BlockHeader) where
 data NewMinedBlock = NewMinedBlock
     { _minedBlockHeader :: ObjectEncoded BlockHeader
     , _minedBlockTrans :: Int
-    , _minedBlockSize :: Int64 }  -- ^ Bytes
+    , _minedBlockSize :: Int }  -- ^ Bytes
     deriving (Eq, Show, Generic)
     deriving anyclass (ToJSON, NFData)
